+++ mall-admin/src/main/java/com/zzy/malladmin/service/impl/RedisServiceImpl.java
@@ -3,7 +3,6 @@ package com.zzy.malladmin.service.impl;
 import com.zzy.malladmin.service.RedisService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Map;
@@ -70,7 +69,7 @@ public class RedisServiceImpl implements RedisService {
     }
 
     @Override
-    public Long incr(String key, Long delta) {
+    public Long incr(String key, int delta) {
         return redisTemplate.opsForValue().increment(key, delta);
     }
